Se reorganizo los componentes por carpetas para su mantenibilidad y se agrego de nuevo el responsivo para los iconos de inicio y espacio para tus eventos
