+++ src/components/events/Lugar/Galeria.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const Galeria = () => {
+  return (
+    <div>Galeria</div>
+  )
+}
