+++ src/index.css
@@ -10,12 +10,13 @@ h2.welcome {
 
 .custom-carousel-section {
   text-align: center;
-  padding: 5px;
+  padding: 10px;
   background-color: #ffffff; /* Fondo blanco */
   border-radius: 25px; /* Esquinas redondeadas */
   box-shadow: 0 1px 1px rgba(0, 0, 0, 0.1); /* Sombra para dar profundidad */
   border: none; /* 0 solid transparent;  Borde inicial transparente */
-  margin: 10px auto; /* Margen superior e inferior para espacio entre secciones */
+  margin: 20px auto; /* Margen superior e inferior para espacio entre secciones */
+  border: 2px solid #FCE300;
 }
 
 .custom-carousel-link {
@@ -79,7 +80,7 @@ h2.welcome {
 }
 
 .custom-carousel-image {
-  width: 98%;
+  width: 100%;
   height: 250px;
   object-fit: cover;
   border-radius: 10px 10px 0 0;
@@ -134,18 +135,25 @@ h2.welcome {
   font-weight: bold; /* Estilo para el texto de la capacidad */
 }
 
+.no-elements-message {
+  text-align: center; /* Centra el texto */
+  font-size: 1.2em; /* Tamaño de fuente */
+  color: #DC143C; /* Color carmesí */
+  margin: 20px 0; /* Espaciado superior e inferior */
+  font-weight: bold; /* Negrita */
+}
 
 
 
 /* Consultas de medios para pantallas más pequeñas */
 @media (max-width: 768px) {
+
   h2.welcome {
     text-align: left;
     font-size: 1em;
     margin-top: 20px;
     margin-bottom: -15px;
   }
-
 
   .custom-carousel-title, .icon-container {
     font-size: 1em;
@@ -200,6 +208,8 @@ h2.welcome {
   .detail-icon {
     margin-right: 6px; /* Menor espacio en pantallas pequeñas */
   }
+
+  
 }
 /****************************  Colores para los iconos de la pagina principal        ***********************/
 
@@ -286,8 +296,18 @@ h2.welcome {
   flex: 1; /* Permite que el texto ocupe el espacio disponible */
 }
 
-
-
 .custom-carousel-info h2{
   padding: 0px;
+}
+
+
+@media (max-width: 768px) {
+  .detail-icon.celebration{
+    font-size: 15px;
+  }
+
+  .detail-info{
+    font-size: .9em;
+    }
+    
 }
