+++ src/routing/AppRouter.js
@@ -1,18 +1,18 @@
 import React from 'react'
 import { BrowserRouter, Routes, Route} from 'react-router-dom'
-import { Inicio } from '../components/general/Inicio'
+import { Header } from '../components/common/Header'
+import { Footer } from '../components/common/Footer'
+import { Inicio } from '../components/home/Inicio'
 import { SobreNosotros } from '../components/general/SobreNosotros'
 import { ReservaEvento } from '../components/general/ReservaEvento'
 import { EspaciosEventos } from '../components/general/EspaciosEventos'
 import { Galeria } from '../components/general/Galeria'
-import { Jardines } from '../components/events/Jardines/Jardines'
-import { Header } from '../components/common/Header'
-import { Footer } from '../components/common/Footer'
-import { InfatilesList } from '../components/events/Infantiles/InfatilesList'
-import { SalonesList } from '../components/events/Salones/SalonesList'
-import { TerrazasList } from '../components/events/Terrazas/TerrazasList'
-import { HaciendasList } from '../components/events/Haciendas/HaciendasList'
-import { DetalleLugar } from '../components/general/DetalleLugar'
+import { Jardines } from '../components/events/TiposEventos/Jardines/Jardines'
+import { InfatilesList } from '../components/events/TiposEventos/Infantiles/InfatilesList'
+import { SalonesList } from '../components/events/TiposEventos/Salones/SalonesList'
+import { TerrazasList } from '../components/events/TiposEventos/Terrazas/TerrazasList'
+import { HaciendasList } from '../components/events/TiposEventos/Haciendas/HaciendasList'
+import { DetalleLugar } from '../components/events/Lugar/DetalleLugar'
