+++ src/styles/general/typeEventCard.css
@@ -8,6 +8,7 @@ Muestra las paginas de espacios para tu evento / Jardin/infantiles/salon/etc ***
   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); /* Sombra para profundidad */
   margin: 20px auto; /* Espacio arriba y abajo, y centrado */
   text-align: center; /* Centra el texto */
+  border: 2px solid #A0522D;
 }
 
 .events-detail{
@@ -108,6 +109,7 @@ Muestra las paginas de espacios para tu evento / Jardin/infantiles/salon/etc ***
     border-radius: 10px;
     box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
     overflow: hidden;
+    border: 2px solid #FCE300;
   }
   /* Imágenes */
   .jardin-image,
@@ -199,15 +201,15 @@ Muestra las paginas de espacios para tu evento / Jardin/infantiles/salon/etc ***
 }
 
 @media (max-width: 768px){
-
+   
+  .detail-item span {
+    padding-top:10px;
+  }
+  
   .detail-info {
     padding: 0px;
   }
 
-  .detail-item span {
-    padding-top:0px;
-  }
-  
   .jardin-cards-container,
   .infantil-cards-container,
   .salon-cards-container,
